add a simple filetype detector instead of relying on file extension
just a bit of "magic" :)

Flac are easy, they always start with "fLaC".  mp3 are weird because
they either start with "ID3" (but this theoretically only ensures that's
a tagged file, not an mp3) or 0xFF 0xFB.

Ogg Opus should have a magic sequence "OpusHead" somewhere near the
start of the file but also have the ogg' "OggS" magic bytes.  I hope
it's enough to distinguish between Ogg Opus and Vorbis.

Another option would be to refactor play_oggvorbis/opus to not close the
file on failure and try in cascade the play_* functions, but it's more
complex and this solution seems to be enough.