don't set the imsg fd as blocking mode
just do a poll before imsg_read in player_disptach to wait for data to
read.  It's not performance-critical code, so this is fine.

If it were important not to do an extra poll(2), for example when we're
called is called after player_pendingimsg we already know there's
something to read, we could move the polling in the case `n == 0' below.