consume all enqueued messages before calling imsg_read
player_dispatch reads only one imsg from the ibuf.  Next time it's
called, the other messages on the ibuf (if any) are discarded and new
ones are read.  This can cause the player process to go out-of-sync with
the main process if multiple messages were "bundled" in the same chunk.

To avoid this, always try to imsg_read before.  If it succeedes, we read
process the equeued messages one by one, if it fails then we poll for
new data and call imsg_read to process them and retry.

This fixes a bug where amused could be "confused" by running

	$ amused pause ; amused stop ; amused play

in a loop a few times.  This bug and the repro were reported two months
ago by Dirk-Wilhelm Peters, thanks!  (and sorry it took so long to
understand and fix the issue)