free the FILE on op_open_callbacks failure
the API is a bit of a shitshow.  op_fdopen is just a wrapper around
FILE*, but returns void* for don't know what Windows limitation.

if op_open_callbacks fails we have to free the stream by ourselves.
the documentation implies that this void* pointer is, in fact, a
FILE*.