reset path buffer before writing to it
otherwise path, which is filled by the garbage that's on the stack, may
not end with a NUL byte.  main_enqueue fails with "malformed data" if
the path sent doesn't end with a NUL.

(i.e. the fact that in every case the path is NUL-terminated is not
relevant.  For semplicity, I'm working with paths long PATH_MAX bytes
and only checking that path[PATH_MAX-1] is NUL when handling the imsgs.)

Issue reported by Dirk-Wilhelm Peters, thanks!