don't loop indefinitely in ctl_connect
if the server fails to start for wathever reason, we end up looping in
ctl_connect indefinitely as we try to spawn the deamon (that dies) and
try to connect to it.

Add an arbitrary maximum number of retries before giving up.

Reported by sikmir in the GitHub issue #1, thanks!