refactor control_notify
it's awkward to call it with the address of the imsgev struct inside
the ctl_conn struct.  Just relay the info to every client in monitor
mode.  After all, if a client is in monitor mode *and* issues
commands, it's not strange that they're echoed back (note that this is
currently impossible, the command line client can either issue
commands or be in monitor mode.)