monitor: fix reported event
rethink a bit which event are reported and when.

- drop 'toggle' event as it's not useful; now 'amused toggle' will
report the play/pause event.

- replace 'flush' with 'load'.  there's no real difference between a
flush and an empty load.  (also, less event the better)

- some events (next/prev/jump) are reported earlier to avoid possible
issues (when messing around with the playing queue we can end up in
every possible state.)

- report 'seek' only after it really happened (i.e. after the player
process sent the new position.)  There's still a possible race here
(we can receive a previous IMSG_POS and think it's the reply to the
seek) but it it's hard/impossible to work around.

- drop 'restart'.  we have 'seek' now which is better (and actually
restart is implemented on top of seek.)

- add 'seek' to the default list of events.