alsa: fix snd_pcm_writei usage
snd_pcm_writei works in terms of audio frames, not buffer (bytes)
length like snd_write does, so account for that.  To do so, prepare
a `bpf' (bytes per frame) multiplier in the configure stage and use
it to convert back and forward between the buffer length and the
number of frames.

While here, also use snd_pcm_avail_update() to know how many frames
may be written.