amused-web: make the filtering work
Generate the HTML for the filter input in js (it's useless to have
it there otherwise) and actually filter the entries.  It's a basic
case insensitive match, but it's a start.  To do so, add dispatch
event type X: to mark the end of the playlist so js can replay the
filtering.