add an libao audio backend
This backend uses libao to play music.  It's a bit convoluted since
libao doesn't provide an async interface, so we have to run it in
a separate thread.  Then, there's some notification via a shared
socketpair because the main loop is around poll().

This actually doesn't work OOTB on OpenBSD due to a restrictive
pledge(), but it's not a issue since sndio should be used there.
libao is the last resort.