add a liboboe audio backend for android
Oboe is a Google C++ library for audio on Android.  The backend is
currently based on the libao skeleton since I couldn't get it to
play audio in a non-blocking way.  (It would also be pointless since
there isn't a way to poll(1).)

It would be worthing experimenting with the callback API.

So far, it works on my phone under termux.  I can control amused
with amused-web.

It still lacks a test in the configure since we would need to use
C++ for it.