add mpris2 support
MPRIS2 is a "standard" to control music player via DBus.  This is done
via a separate amused client, called `amused-mpris2', that connects to
DBus and proxies there the events and converts method calls to amused'
commands.

It still has some issues, playerctl doesn't seem to like it much, but
it's enough to have multimedia keys and bluetooth headsets support in
KDE working.

It is disabled by default, to build it pass --with-mpris2 to the
configure script.  Has a dependency on glib.