add `shuffle' subcommand
Funnily enough, before 0.1, it was the will to not implement commands
like shuffle or sort that made me realize I could use show | load as a
pattern to transform the playing queue.

However, edd@ come up with a particular request for `shuffle', that is
to shuffle only the rest of the playing queue.  It's difficult to
script, but very easy to implement.

While here, also add -a to shuffle all the playing queue.